Split the MbUnit Core and Framework.
Integrated ILMerge by creating a fake C# project with custom targets to run ILMerge.  It also fools Visual Studio into choosing the right project references based on the current configuration.  In the 'Debug' configuration, ILMerge isn't used and MbUnit.Gallio.dll is just a stub assembly.  In the 'Release' configuration, we use ILMerge and MbUnit.Gallio.dll is the merged Framework, Core and dependencies.

Now... of course I've just decided to change how Framework and Core are defined so ILMerge will go away.  I'm checking this in to shelve it on a branch then I'll be reworking all of this stuff again.  *sigh*