Refactored the test runner to clean things up. Introduced ITestRunner as a common (mockable) abstraction for a test runner and a BaseRunner that captures most of the lifecycle but leaves a number of details open.  Runners also implement IDisposable now like they should (since they own ITestDomains).
AutoRunner implements ITestRunner and takes an ITestDomainFactory parameter
to specify how it creates its test domain.  A simple static method takes
care of setting up the runtime and everything needed in the common case
where the default test domain is being used (currently IsolatedTestDomain).

Added a RuntimeSetup object to provide essential configuration parameters
to a runtime.  Currently it specifies additional plugin directories.
Created a new ICoreRuntime interface to capture core-visible facets of the
runtime that the framework shouldn't see when it looks at IRuntime.

Added IIsolatedTestDomainContributor to capture plug-in contributions to
IsolatedTestDomain such as adding extra assembly binding redirects if needed.

Added a command-line option to Echo for specifying additional plugin directories.
This option is particularly useful for self-hosting unit tests without needing
to copy all of the plugins from their various Visual Studio bin folders.

Significantly refactores the template tree and provided common interfaces
for the template and test tree.  In particular, the templates and tests
are considered nodes in a plain old tree with a list of children and all
that.  I removed the nonsense of having some templates that might not support
AddChild.  They're just plain nodes now.

Implemented the notion of an ITemplateBinding that captures the identity
of a template whose parameters have been bound to particular values in a
particular test scope.

Added classes to represent the RootTest, RootTemplate and
RootTemplateBinding so we can capture any special behavior cohesively.

Implemented enough of the test tree building process to start to see
interesting behavior emerge (and to stress the underlying abstractions...
which prompted further refactoring).

Added initial support for enumerating tests in an MbUnit v2 test assembly using
the MbUnit v2 FixtureExplorer.

What's cool is that Echo can enumerate and dump the template and test trees
of Gallio and MbUnit v2 test assemblies.  It's still a bit primitive (missing
some metadata and robust error handling -- see TODOs) but it seems to work
fine.  This means we can provide Icarus with real test data to display...