Added Filters for all of the basic cases involving assemblies, types, members, etc... as well as combinators.
Added Options classes for providing an extensible set of parameters to the test harness during template/test enumeration and test execution without having to update each call-site to pass them around.  The plumbing is in place but I'm not making use of it quite yet.  (Oh no!  Bad TDD... but then I haven't really applied TDD to any of this stuff.)