More progress towards being able to run MbUnit Gallio tests.
Implemented [Test], [TestFixture], [TestFixtureSetUp], [TestFixtureTearDown],
[SetUp] and [TearDown].

Fleshed out more of the first Gallio integration test.  It fails of course
(which is good since we're missing key pieces!).

Added code to run the tests.  Of course it's not complete yet since we need
to be able to set up the current IContext so that we can grab execution
log output and capture assertions.

Added an IsGenerator property to ITemplate.  The idea is that templates
that are classified as generative are automatically bound to whatever
values happen to be available in the parent template's scope.  Thus [Test]
and [TestFixture] both declare generative templates.  Non-generative templates
are ignored during this process and must be bound explicitly, if at all.
Thus [SetUp], [TearDown] and others all declare non-generative templates.

Renamed MbUnitFixtureTemplate to MbUnitTestTemplate, TestPatternAttribute
to MethodPatternAttribute and FixturePatternAttribute to TypePatternAttribute.
The more general names are useful for signalling that the templates aren't
always generating tests and fixtures, they could be doing other things too.
Creates a new TestPatternAttribute and FixturePatternAttribute subclass that
pretty much just sets IsGenerative to true.  Improved the XML documentation
comments somewhat and a swatted a few silly bugs in currently untested
parts of the framework.

Performed a first pass code sweep for localizable strings in the MbUnit 
plugins and in portions of the framework.  RGreatEx is being very helpful for this.