Reconsidered how a few custom tools are installed.
Namely, it's nice to have everything work out of the box if you just run "Build.bat".  On the other hand, it'd be unreasonable to try to do that with "Release.bat" because it depends on too many tools (including the Visual Studio 2005 SDK!)

So I've moved some of the tools to a "Setup Files" folder.

Enhanced the build script to check for dependencies and fixed a small bug.