Refactored and enhanced the console support.
The console logger and progress monitor can now co-exist peacefully so
we can display real-time results to the console.  Added the /echo-results 
option to MbUnit.Echo to enable this kind of display.

Automatically disable progress monitoring when output is being redirected.

Now the build script is self-hosting on MbUnit.Echo!  There are bugs to
be fixed here...  Some tests fail when they shouldn't.

Added a FullName property to steps to get rid of the hacky formatting that
was happening in a handful of places.

Fixed the HTML report template to correctly read results from step runs.

Started externalizing strings from the MbUnit.Core assembly.

Added behavior to ReportMonitor to easily get consolidated results from
the end of a step run.  We are now using this feature in the new LogMonitor
and in the TDNetLogMonitor.

Replaced LogMonitor with a new class that can echo test results back to
an ILogger.  That's used to support the new /echo-results argument.
Naturally it could be used in the various build tasks as well if we like.

Eliminated hard-coded paths to the test samples.  It's better if the
test cases just reference the sample assembly instead.  That way its path
is readily obtained and the dependency is made explicit.

Misc. bug fixes.