Committed some proof of concept code using Mono.Cecil to perform runtime interception in two different ways:
1. By statically instrumenting the assembly to include interception code that can be hotwired on demand.  (easy)
2. By loading the assembly into the dynamic assembly context and patching method bodies on demand using MethodRental.  (hard)

There are lots of bugs and missing pieces but it demonstrates the principle.

Unfortunately I ran into two Mono.Cecil bugs along the way:
- http://bugzilla.ximian.com/show_bug.cgi?id=82814
- http://bugzilla.ximian.com/show_bug.cgi?id=82820

Ultimately the idea is for MbUnit to expose an interception facility that can be used by client code for deep tracing and mocking.  Generally the profiler API is better at this sort of thing but there are other interesting scenarios, particularly if the feature is generalized beyond MbUnit.

I'm not entirely certain how useful this is so I don't plan on doing much with this code anytime soon.  I just couldn't get it out of my mind...  Now I can get some peace in my brain to work on more important things.