Moved the Report XSL and resources to a new plugin called MbUnit.Plugin.Report where they are now just plain content files.  This allows them to be edited by end-users and enables easier integration of new report formats.
Upgraded the Castle binaries to 1.0 RC3.

Expanded on the IFileSystem abstraction a bit and used it to improve the testability of the report load/save and formatting API.

Fixed a subtle bug with assembly resolution by introducing explicit assembly binding qualifications for bootstrap assemblies.  It was missing earlier and causing problems because it was resolving Castle assemblies in the GAC if it couldn't find them in the current directory when bootstrapping.  Ugh!

Simplified the installer rules a tad and added the new Reports plugin.  Also added in Icarus for good measure.  I'd left it out previously because there wasn't anything interesting to see at the time.

Added XHTML, HTML and CCNET variants of the (X)HTML report by using stylesheet includes.

Fixed XHTML and HTML validation issues.  We're now fully HTML 4.01 strict and XHTML 1.0 strict compliant!

Other minor tweaks and fixes.