Added a host application for Gallio which can be used to launch tests out of process.
Refactored the test domain code to extract a new primitive: IHost.
A host is simply an abstraction around an remote execution environment which can perform whatever services Gallio requires.  There are currently two implementations of IHost: one based around an isolated app-domain within the current process, and another that launches the Gallio.Host.exe application to create an out-of-process environment.
The HostTestDomain replaces IsolatedTestDomain and is designed to work with any suitable host.

Replaced IIsolatedTestDomainContributor with IHostTestDomainContributor.  The contribution now applies to a HostSetup object which carries configuration settings for the host.  This feature now decouples the test domain from the details of creating a suitable Xml configuration file for the host environment runtime.  I've also added support for a number of features that weren't previously accessible.  This work paves the way to allowing test assembly *.dll.config files to be loaded and used while at the same time overlaying the configuration information with critical parameters necessary to provide a stable hosting environment.

Added an abstraction for .Net remoting channels.

Moved the contents of the Core namespace to the Hosting namespace where I think they make a bit more sense.

Enabled the use of SGen to pre-generate Xml serializers for the Gallio assembly.  Note that this required hacking the project file a little bit because as written in .Net 2.0, SGen only gets used for web service proxy types.

Fixed a bug in the MSBuild task related to displaying all of the exception details when the runner dies on a fatal exception.

Fixed a bug in the CommandLineArgumentParser when displaying help when there are no default arguments.

Extracted a new framework type called ConsoleProgram to assist with the implementation of robust command-line programs that interact with the console.