Added Unresolved member wrappers that enhance the compatibility of the reflection layer with native reflection by providing an alternative view of the adapters that can be used by code that expects to deal with native types.  A typical use of these wrappers is to support the instantiation of Attributes that have arguments of type Type.  Previously attempting to initialize the attribute would fail and throw an error in environments like the R# test runner.  This limitation caused attributes like [TestsOn] to generate errors unnecessarily.  It takes a little work, but we can subclass the native reflection member objects like Type, MethodInfo, etc. to provide better support for code that just has to use the native type instead of the adapter type.
Added a throwOnError parameter to the Resolve() method of IMemberInfo to control whether failure to resolve a member should throw an exception or return an Unresolved member wrapper instead.

Added a few new properties to the reflection adapters that cover missing functionality such as generic type definitions and getting the attributes associated with a method return value.

Tweaked the R# test runner output to show the stream names and all contents.

Rearranged the order in which Echo prints usage information to make the application appear a little more responsive by avoiding a long wait on initialization.