Refactored SlotBinder into ObjectCreationSpec and MethodInvocationSpec which now cleanly separate the initial task of resolving members and arguments from that of actually instantiating the object or invoking the method.
Moved the point of data conversion into the ObjectCreationSpec and MethodInvocationSpec objects.  This was necessary because it was not possible to convert arguments ahead of for generic methods when the parameter was a generic method parameter until we had bound a value to the generic method parameter.  Consequently DatabindingContext no longer performs any data conversions itself.  They are all deferred until we create the Spec objects at which point we will have all of the necessary information in one place.
Removed the ValueType property of DataBinding since conversions are being handled differently now.  This has produced some useful simplifications in numerous places.

Added a new formatting rule for Types to improve presentation instead of relying on the default ToString wrapped up in { }.
Improved the conversion to string formatting rule to use an unquoted string literal format to avoid weird formatting glitches.

Split the IPatternTestHandler into two to create an IPatternTestInstanceHandler with the test instance actions.  Likewise split up PatternTestActions.
Added a DecorateTestInstance phase to support per-instance decorations.
Added a FixtureInitializerAttribute to MbUnit.
Added a TestDecoratorAttribute to MbUnit as a convenience for implementing custom test decorators.

Added support for comparing NativeMemberWrappers for equality.
Note: Had to workaround a .Net framework limitation (bug?) involving generic type instantiation information being lost in certain cases.  Refer to NativeMemberWrapper for more details in the comments.