Work in progress.
Redesigned the ITestRunner abstraction and how it dispatches events and is extended.  The overall story is somewhat simpler now.
The generation of the Report object is now built into the test runner itself.
We will also support registration of standalone extensions from the command-line similarly to the MSBuild logger extension mechanism.
Renamed some of the core operations.
Added code to eliminate unnecessary copying of the Gallio.Host.exe program when a default configuration is used.
This checkin paves the way for some bigger changes in terms of how we manage test domains overall so that we can run test assemblies in separate domains and support assembly configuration files.