Significantly reworked documentation on all MbUnit attributes and asserts and repaired many inconsistencies. Optimized the sample test runner to eliminate redundant and expensive sample execution. Moved timeout functionality into Sandbox. Added extra tests for Sandbox and StructuredTestLog components. Added a new feature for building static and dynamic test suites. Fixed some bugs handling dynamic tests.
Breaking changes:

Modified the meaning of the Local test runner to really be local again by adding a ShareAppDomain property to HostedTestDriver.
Eliminated the redundantly created hosting AppDomain of the AppDomain test runner.
Modified the handling of metadata in test steps so that the primary step normally inherits the metadata of its test.  Updated the reports accordingly.
Changed test step API to support timeouts.