Fixed some bugs in the CLI PE file reading and added support for getting useful bits of metadata info into an AssemblyMetadata object including the processor architecture and CLI major/minor version.
I was originally going to auto-detect the necessary .Net framework version using the CLI major/minor version from the PE file but instead just opted to default to .Net 4.0 when installed and fallback on .Net 2.0 otherwise.  Auto-detection is a bad plan because we might try to run tests that don't live in a compiled assembly but that do require a particular runtime version anyways.