Added support for displaying embedded HTML, XHTML, Text and XML attachments by using a hidden IFrame to load the content.  This works fine in FireFox but in IE it will only work when a report is viewed online.
More on IE:

We use the IE Mark of the Web (MOTW) token to enable scripts to execute within HTML documents in the Local Machine Zone.  However, the MOTW expresses a transitive constraints on all linked documents and files to also have an MOTW.  Consequently when the user clicks on a link to an attachment (which does not have an MOTW), the browser simply ignores the request.  Likewise when we try to point an iframe at the contents of attachment the navigation fails.

One possible workaround is to remove the MOTW.  In that case IE displays a security warning banner.  When the user accepts the banner then everything works great!  Unfortunately this does not yield a satisfactory viewing experience for local reports.

Another option is to set some registry keys to disable the Local Machine Lockdown policy.  This seems rather impolite and risky it will expose the user's machine to other local scripting vulnerabilities.

I have attempted a few approaches at mixing MOTW and non-MOTW content but the sandbox has so far resisted my attempts to escape it.  It is certainly possible to achieve this by leveraging an ActiveX or a custom protocol handler such as the "gallio:" protocol we use to navigate to source.  The question is whether it can be made relatively secure and robust.