Removing PingMessage.
When the TestIsolationClient side (where the tests are executing) is being debugged it may not send the PingMessage frequently enough because the process is halted. TestIsolationServer doesn't distinguish between this and a hung process and aborts the test run. I'm not aware of a way to detect this so I'm removing the PingMessage mechanism. This means that debugging tests will no longer cause a time-out with the trade-off that Gallio will not be able to detect a client that for some reason is no longer sending pings. If the process the client is running in crashes the server will still be able to detect that because the pipe will die as well.

AFAIK TestIsolationClient/TestIsolationServer are only used by the AutoCAD runner.